change getInt to getDouble

+++ app/src/main/java/edu/agh/mapnote_androidapp/DbHelper.java
@@ -90,8 +90,8 @@ public class DbHelper extends SQLiteOpenHelper {
             do{
                 Note note = new Note();
                 note.setId(cursor.getInt(0));
-                note.setLatitude(cursor.getInt(1));
-                note.setLongitude(cursor.getInt(2));
+                note.setLatitude(cursor.getDouble(1));
+                note.setLongitude(cursor.getDouble(2));
                 note.setAddress(cursor.getString(3));
                 note.setDate(cursor.getString(4));
                 note.setNoteContent(cursor.getString(5));
